requirements.txt
aiohappyeyeballs==2.3.6
aiohttp
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==25.3.0
axios==0.4.0
billiard==4.2.0
black==24.10.0
blinker==1.9.0
cachetools==5.5.2
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.1
datasets==3.5.0
decouple==0.0.7
dill==0.3.8
distlib==0.3.9
distro==1.9.0
Django==5.1.5
django-allauth==65.0.2
django-cors-headers==4.4.0
django-crispy-forms==2.1
django-environ==0.11.2
django-fernet-fields==0.6
django-formtools==2.5.1
django-gmailapi-backend==0.3.2
django-otp==1.5.4
django-phonenumber-field==7.3.0
django-two-factor-auth==1.15.5
django_csp==3.8
fastapi==0.115.4
filelock==3.18.0
flake8==7.0.0
Flask==3.0.1
frozenlist==1.4.1
fsspec==2024.3.1
git-filter-repo==2.45.0
google-api-core==2.23.0
google-api-python-client==2.139.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gradio_client==1.3.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.23.3
idna==3.7
intel-openmp
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jwt==1.3.1
kombu==5.3.7
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mkl
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.3
oauthlib==3.2.2
openai==0.28.0
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==6.30.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
pydantic
pydantic_core
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.2.3
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.1
qrcode==7.4.2
redis==5.2.0
regex==2024.9.1
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.1
rsa==4.9
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.0
starlette==0.41.2
sympy
tbb==2022.0.0
textblob==0.17.1
tokenizers
torch
tqdm==4.67.1
transformers
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
virtualenv==20.26.6
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.6
xxhash==3.4.1
yarl==1.9.4